added authorization for api requests

+++ frontend/src/components/workout/Workout.js
@@ -1,15 +1,24 @@
 import './Workout.css';
 import { useWorkoutsContext } from '../../hooks/useWorkoutsContext';
+import { useAuthContext } from '../../hooks/useAuthContext';
 
 //date fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow';
 
 const Workout = ({workout}) => {
     const { dispatch } = useWorkoutsContext();
+    const { user } = useAuthContext();
 
     const handleClick = async () => {
+        if(!user) {
+            return;
+        }
+        
         const response = await fetch(`/api/workouts/${workout._id}`, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         });
 
         const data = await response.json();
